+++ shenyu-integrated-test/shenyu-integrated-test-combination/script/services.list
@@ -16,6 +16,7 @@
 
 http://localhost:9095/actuator/health
 http://localhost:8189/test/path/123?name=tom
+http://localhost:8190/actuator/health
 http://localhost:8011/actuator/health
 http://localhost:55290/test/grpc/hello
 http://localhost:8081/actuator/health
