+++ shenyu-integrated-test/shenyu-integrated-test-http/script/services.list
@@ -16,6 +16,5 @@
 
 http://localhost:9095/actuator/health
 http://localhost:8189/test/path/123?name=tom
-http://localhost:8190/actuator/health
 http://localhost:9195/actuator/health
 http://localhost:9200/_cat/health?v
