+++ shenyu-integrated-test/shenyu-integrated-test-http/src/test/java/org/apache/shenyu/integrated/test/http/combination/ContextPathPluginTest.java
@@ -37,9 +37,6 @@ import java.util.Map;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 
 public final class ContextPathPluginTest extends AbstractPluginDataInit {
 
@@ -62,64 +59,31 @@ public final class ContextPathPluginTest extends AbstractPluginDataInit {
         String pluginResult = initPlugin(PluginEnum.CONTEXT_PATH.getName(), "");
         assertThat(pluginResult, CoreMatchers.is("success"));
         final String ruleHandle = "{\"contextPath\":\"\", \"addPrefix\":\"/error\"}";
-        String message = initSelectorAndRules(PluginEnum.CONTEXT_PATH.getName(), "",
-                buildSelectorConditionList("/http/**"), buildRuleLocalDataList("/http/**", ruleHandle));
+        String message = initSelectorAndRules(PluginEnum.CONTEXT_PATH.getName(), "", buildSelectorConditionList(), buildRuleLocalDataList(ruleHandle));
         assertThat(message, is("success"));
     }
 
-    private static List<ConditionData> buildSelectorConditionList(final String paramValue) {
+    private static List<ConditionData> buildSelectorConditionList() {
         ConditionData conditionData = new ConditionData();
         conditionData.setParamType(ParamTypeEnum.URI.getName());
         conditionData.setOperator(OperatorEnum.MATCH.getAlias());
-        conditionData.setParamValue(paramValue);
+        conditionData.setParamValue("/http/**");
         return Collections.singletonList(conditionData);
     }
 
-    private static List<RuleLocalData> buildRuleLocalDataList(final String paramValue, final String ruleHandleString) {
+    private static List<RuleLocalData> buildRuleLocalDataList(final String ruleHandleString) {
         RuleLocalData ruleLocalData = new LocalPluginController.RuleLocalData();
         ruleLocalData.setRuleHandler(ruleHandleString);
-        ruleLocalData.setConditionDataList(Collections.singletonList(buildConditionData(paramValue)));
+        ruleLocalData.setConditionDataList(Collections.singletonList(buildConditionData()));
         return Collections.singletonList(ruleLocalData);
     }
 
-    private static ConditionData buildConditionData(final String paramValue) {
+    private static ConditionData buildConditionData() {
         ConditionData conditionData = new ConditionData();
         conditionData.setParamType(ParamTypeEnum.URI.getName());
         conditionData.setOperator(OperatorEnum.MATCH.getAlias());
-        conditionData.setParamValue(paramValue);
+        conditionData.setParamValue("/http/**");
         return conditionData;
-    }
-    
-    @Test
-    public void testRewriteContextPath() throws IOException {
-        Map<String, Object> result = HttpHelper.INSTANCE.getFromGateway("/order/order/findById?id=3", Map.class);
-        assertNotNull(result);
-        assertNull(result.get("id"));
-        assertNull(result.get("name"));
-        
-        result = HttpHelper.INSTANCE.getFromGateway("/http/order/findById?id=3", Map.class);
-        assertNotNull(result);
-        assertEquals("3", result.get("id"));
-        assertEquals("hello world findById", result.get("name"));
-        
-        // rewrite context path from /order to /http
-        String cleanMessage = cleanPluginData(PluginEnum.CONTEXT_PATH.getName());
-        assertThat(cleanMessage, is("success"));
-        setupRewriteConfiguration();
-        
-        Map<String, Object> response = HttpHelper.INSTANCE.getFromGateway("/order/order/findById?id=4", Map.class);
-        assertNotNull(response);
-        assertEquals("4", response.get("id"));
-        assertEquals("hello world findById", response.get("name"));
-    }
-    
-    private void setupRewriteConfiguration() throws IOException {
-        String pluginResult = initPlugin(PluginEnum.CONTEXT_PATH.getName(), "");
-        assertThat(pluginResult, CoreMatchers.is("success"));
-        final String ruleHandle = "{\"contextPath\":\"/order\", \"addPrefix\":\"\", \"rewriteContextPath\":\"/http\", \"percentage\":100}";
-        String message = initSelectorAndRules(PluginEnum.CONTEXT_PATH.getName(), "",
-                buildSelectorConditionList("/order/**"), buildRuleLocalDataList("/order/**", ruleHandle));
-        assertThat(message, is("success"));
     }
 
     @AfterAll
@@ -132,8 +96,7 @@ public final class ContextPathPluginTest extends AbstractPluginDataInit {
         String pluginResult = initPlugin(PluginEnum.CONTEXT_PATH.getName(), "");
         assertThat(pluginResult, CoreMatchers.is("success"));
         final String ruleHandle = "{\"contextPath\":\"/http\"}";
-        String message = initSelectorAndRules(PluginEnum.CONTEXT_PATH.getName(), "",
-                buildSelectorConditionList("/http/**"), buildRuleLocalDataList("/http/**", ruleHandle));
+        String message = initSelectorAndRules(PluginEnum.CONTEXT_PATH.getName(), "", buildSelectorConditionList(), buildRuleLocalDataList(ruleHandle));
         assertThat(message, is("success"));
     }
 }
